Specimens/new.py
# Generated at 2024-06-18 01:17:05.773799
# Query:
# There was a problem running the code: ```python
# import random
# 
# def random_number_generator():
#     return [i for i in range(1,7)]
# 
# @pytest.mark.parametrize("expected_count", [(1000, 2000), (3000, 4000), (5000, 6000)])
# def test_random_number_probability(expected_count):
#     generator = iter(random.randint(1,6) for _ in range(10000))
#     count_dict = {}
# 
#     for _ in range(10000): 
#         num = next(generator)
#         if num not in count_dict:
#             count_dict[num] = 1
#         else:
#             count_dict[num] += 1
# 
#     for expected, actual in zip(expected_count, list(count_dict.values())):
#         assert abs(actual - (10000 / 6) * expected) < 50 
# 
# @pytest.mark.parametrize("expected_range", [(1,6)])
# def test_random_number_range(expected_range):
#     generator = iter(random.randint(1,6) for _ in range(10))
#     numbers = [next(generator) for _ in range(10)]
#     assert all(exp <= num <= sup for num, (exp, sup) in zip(numbers, [(expected_range[0], expected_range[1])] * 10))
# ```
# 
# The `error was: Traceback (most recent call last):
#   File "C:\Users\kwokd\Documents\1_PROJECTS\ai_game\Specimens\new.py", line 51, in <module>
#     @pytest.mark.parametrize("expected_count", [(1000, 2000), (3000, 4000), (5000, 6000)])
#      ^^^^^^
# NameError: name 'pytest' is not defined. Did you mean: 'bytes'?
# 
# 
# Please try again
#
# python
import pytest
import random

# Define the function to generate a list of random numbers between 1 and 6
def random_number_generator():
    return [i for i in range(1,7)]

@pytest.mark.parametrize("expected_count", [(1000, 2000), (3000, 4000), (5000, 6000)])
def test_random_number_probability(expected_count):
    # Initialize a generator to produce random numbers between 1 and 6
    generator = iter(random.randint(1,6) for _ in range(10000))
    count_dict = {}

    # Count the occurrences of each number
    for _ in range(10000): 
        num  = next(generator)
        if num not in count_dict:
            count_dict[num]  = 1
        else:
            count_dict[num] += 1

    # Check that the counts are close to the expected counts
    for expected, actual in zip(expected_count, list(count_dict.values())):
        assert abs(actual - (10000 / 6) * expected) < 50 

@pytest.mark.parametrize("expected_range", [(1,6)])
def test_random_number_range(expected_range):
    # Initialize a generator to produce random numbers between 1 and 6
    generator = iter(random.randint(1,6) for _ in range(10))
    numbers = [next(generator) for _ in range(10)]

    # Check that all the generated numbers are within the expected range
    assert all(exp <= num <= sup for num, (exp, sup) in zip(numbers, [(expected_range[0], expected_range[1])] * 10))
